Modify Document CRUD

+++ app/Controllers/Http/DocumentController.js
@@ -1,10 +1,18 @@
 'use strict'
 
+const STATUS = {
+  'CREATED': 'created',
+  'SEND': 'send',
+}
+
 /** @typedef {import('@adonisjs/framework/src/Request')} Request */
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
 /** @typedef {import('@adonisjs/framework/src/View')} View */
 
 const Document = use('App/Models/Document')
+const Solicitation = use('App/Models/Solicitation')
+const Question = use('App/Models/Question')
+
 /**
  * Resourceful controller for interacting with documents
  */
@@ -24,12 +32,18 @@ class DocumentController {
    * POST documents
    *
    */
-  async store ({ request, response }) {
-    const data = await request.only(['name','type','status','solicitation_id','created_by'])
-    const document = await Document.create(data)
+  // async store ({ request, response }) {
+  //   const data = await request.only(['name','type','solicitation_id','created_by'])
+  //   const {questions} = await request.only(['questions']);
+  //   //Verifying if there are questions on Document
+  //   if (questions.length){
+  //     const document = await Document.create({...data, status: STATUS.CREATED})
+  //     document.questions().createMany(questions)
+  //     return response.ok(document)
+  //   }
 
-    return document;
-  }
+  //   return response.badRequest({message: "O documento não foi preenchido"})
+  // }
 
   /**
    * Display a single document.
@@ -39,7 +53,7 @@ class DocumentController {
   async show ({ params, request, response, view }) {
     const document = await Document.findOrFail(params.id)
 
-    await document.load('solicitation')
+    // await document.load('solicitation')
     await document.load('attachments')
     await document.load('questions')
 
@@ -53,14 +67,43 @@ class DocumentController {
    *
    */
   async update ({ params, request, response }) {
-    //Some params dont be chenged
-    const data = await request.only(['name','status'])
-    const document = await Document.findOrFail(params.index)
 
-    document.merge(data)
-    await document.save()
+    const document = await Document.findOrFail(params.id)
 
-    return document;
+    if (document.status === STATUS.CREATED){
+      const {name, questions} = await request.all()
+      document.merge({name: name})
+      await document.save()
+
+      //Updating questions
+      await Promise.all(
+        questions.map(qstns => {
+          if (!qstns.id) {
+            return document.questions().create(qstns)
+          }
+          return Question.query()
+            .where(function () {
+              this
+                .where('id', qstns.id)
+                .andWhere('document_id', params.id)
+            })
+            .first()
+            .then(existingQuestion => {
+              console.log('okokoko')
+              delete qstns.id
+              if (existingQuestion) {
+                existingQuestion.merge(qstns)
+                return existingQuestion.save()
+              } else {
+                return document.questions().create(qstns)
+              }
+            })
+        })
+      )
+      return response.ok(document)
+    }
+
+    return response.forbidden({message: "O documento já foi enviado e não pode ser excluído"})
   }
 
   /**
@@ -69,10 +112,14 @@ class DocumentController {
    *
    */
   async destroy ({ params, request, response }) {
-    const document = await Document.findOrFail(params.index)
-    await document.delete();
+    const document = await Document.findOrFail(params.id)
 
-    return document
+    if (document.status === STATUS.CREATED){
+      await document.delete();
+      return response.ok(document)
+    }
+
+    return response.forbidden({message: "O documento já foi enviado e não pode ser excluído"})
   }
 }
